pyproject.toml
[project]
name = "ia-tp1"
version = "0.0.0"
description = "Trabajo práctico para IA, sobre crear un RAG usando LangChain, y crear un chatbot que responda solamente con los conocimientos proporcionados por dicho RAG."
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "langchain (>=1.2.15,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "langchain-google-genai (>=4.2.1,<5.0.0)",
    "langchain-chroma (>=1.1.0,<2.0.0)",
    "chromadb (>=1.5.6,<2.0.0)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)",
    "langchain-classic (>=1.0.3,<2.0.0)",
    "sentence-transformers (>=5.3.0,<6.0.0)",
    "pypdf (>=6.9.2,<7.0.0)",
    "questionary (>=2.1.1,<3.0.0)"
]


[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
